% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{data_munging}
\alias{data_munging}
\title{Data Munging
Given the raw data, return variables used in the analysis. Munging steps:
\itemize{
\item Categorical variables set to factor
\item Compute questionnaire scores
\item Select relevant variables used in the analysis
}}
\usage{
data_munging(raw_data)
}
\arguments{
\item{raw_data}{a tibble object with the \code{"data/data.csv"}.}
}
\value{
a tibble object with the data used in the analysis
}
\description{
Data Munging
Given the raw data, return variables used in the analysis. Munging steps:
\itemize{
\item Categorical variables set to factor
\item Compute questionnaire scores
\item Select relevant variables used in the analysis
}
}
\examples{
targets::tar_load(raw_data)
data_munging(raw_data)

}
