% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_group_summary}
\alias{get_group_summary}
\title{Get BY Group Summary}
\usage{
get_group_summary(my_data, var, group_by)
}
\arguments{
\item{my_data}{a tibble with the data}

\item{var}{unquoted expression indicating the variable of interest.}

\item{group_by}{character vector indicating the variable to group by.}
}
\value{
a tibble with the grouping variables and mean and SD of the desired
var
}
\description{
Get mean and SD of a numeric variable according to group_by
}
\examples{
targets::tar_load(my_data)
get_group_summary(my_data, var = ext_tot, group_by = c("gender", "grade"))

}
