% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\name{get_ssm_score}
\alias{get_ssm_score}
\title{Compute the SSM Score}
\usage{
get_ssm_score(raw_data)
}
\arguments{
\item{raw_data}{a tibble object with the \code{"data/data.csv"}.}
}
\value{
a numeric vector with the total SSM score for each children
}
\description{
Given the raw data, compute the total SSM score for each children
}
\details{
The total score is given by the average of the items. Higher scores indicate
higher levels of security.

To construct the final score, it is necessary to invert the scores of items
1, 3, 4, 9, 10, 13, and 15. That is: 1 --> 4; 2 --> 3; 3 --> 2; 4 --> 1.
}
\examples{
targets::tar_load(raw_data)
get_ssm_score(raw_data)

}
