% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-utils.R
\name{load_glob_env}
\alias{load_glob_env}
\title{Load Targets in Global Environment}
\usage{
load_glob_env(..., store = targets::tar_config_get("store"))
}
\arguments{
\item{...}{parameters passed to \code{targets::tar_load()} function}

\item{store}{the path to the tergets data store}
}
\value{
The cached value of the target
}
\description{
Load targets in Global Environment
}
\examples{
load_glob_env(my_data)
my_data

}
