% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{my_labeller}
\alias{my_labeller}
\title{Get Plot Labels}
\usage{
my_labeller(x)
}
\arguments{
\item{x}{string indicating the variable name}
}
\value{
string indicating the label
}
\description{
Given the variable name get the corresponding label to use in plots
}
\examples{
my_labeller("ssm_tot")
my_labeller("a new name")

}
