% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reverse_item}
\alias{reverse_item}
\title{Reverse Items}
\usage{
reverse_item(x, min_score = 1, max_socore = 4)
}
\arguments{
\item{x}{numeric vector with the item responses.}

\item{min_score}{numeric value indicating the minimum item possible value.}

\item{max_socore}{numeric value indicating the maximum item possible value.}
}
\value{
numeric vector with the reversed item responses.
}
\description{
Given the maximum and minimum scores, reverse the item value.
}
\examples{
targets::tar_load(raw_data)
x <- raw_data$ssm3
reverse_item(x, min_score = 1, max_socore = 4)

}
