% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets-utils.R
\name{tar_load_all}
\alias{tar_load_all}
\title{Load All Targets Objects}
\usage{
tar_load_all(store = targets::tar_config_get("store"))
}
\value{
all cached value of the target
}
\description{
Load all targets objects
}
\examples{
tar_load_all()

}
